photometry: move rotation-period modules into a rotation/ subpackage
Declutter the photometry/ directory: the six rotation_period_*.py modules are now
a photometry/rotation/ subpackage, and the two small modules are folded into
core.py.

  rotation_period_fourier.py       -> rotation/estimator.py
  rotation_period_fourier_core.py  -> rotation/core.py
  rotation_period_jax.py           -> rotation/jax_backend.py   (name avoids shadowing `jax`)
  rotation_period_wrappers.py      -> rotation/wrappers.py
  rotation_period_types.py         -> folded into rotation/core.py
  rotation_period_scoring.py       -> folded into rotation/core.py

core.py now holds three sections: the Fourier math primitives, the public data
tables (RotationPeriodObservations / RotationPeriodResult /
GroupedRotationPeriodResults), and the period-recovery scoring metrics. A new
rotation/__init__.py re-exports the public API, and adam_core.photometry imports
from it, so the public surface (adam_core.photometry.*) is unchanged.

Behavior-preserving move + import rewrites only. mypy --strict clean on the
subpackage; ruff/black/isort clean; docs build clean under -W (the API reference
autogenerates for the new layout); 32 passed / 1 xfailed (Ivar) / 4 deselected.
