+++ src/adam_core/photometry/rotation/jax_backend.py
