+++ src/adam_core/photometry/tests/test_rotation_period_validation.py
@@ -28,14 +28,14 @@ import numpy as np
 import pyarrow as pa
 import pytest
 
-from adam_core.photometry.rotation_period_fourier import estimate_rotation_period
-from adam_core.photometry.rotation_period_scoring import (
+from adam_core.photometry.rotation.core import (
+    RotationPeriodObservations,
     alias_bucket,
     harmonic_adjusted_error_pct,
     relative_error_pct,
     within_tolerance,
 )
-from adam_core.photometry.rotation_period_types import RotationPeriodObservations
+from adam_core.photometry.rotation.estimator import estimate_rotation_period
 from adam_core.time import Timestamp
 
 DATA_DIR = Path(__file__).parent / "data"
